chore: extract popover JSDOM shim into its own module
Pulls the popover/ToggleEvent shim out of setup.ts so the test setup
file reads as a list of "what we're patching" rather than a wall of
shim implementation. The shim is loaded via a side-effecting import,
matching the pattern used by core-js/whatwg-fetch and the existing
ResizeObserver/HTMLDialogElement mocks.
