+++ resources/assets/js/__tests__/setup.ts
@@ -1,5 +1,6 @@
 import vueSnapshotSerializer from 'jest-serializer-vue'
 import { expect, vi } from 'vite-plus/test'
+import './popover-shim'
 
 declare global {
   interface Window {
@@ -51,55 +52,6 @@ HTMLDialogElement.prototype.showModal = vi.fn(function mock(this: HTMLDialogElem
 HTMLDialogElement.prototype.close = vi.fn(function mock(this: HTMLDialogElement) {
   this.open = false
 })
-
-if (typeof globalThis.ToggleEvent === 'undefined') {
-  class ToggleEventShim extends Event {
-    oldState: string
-    newState: string
-    constructor(type: string, init: EventInit & { oldState?: string; newState?: string } = {}) {
-      super(type, init)
-      this.oldState = init.oldState ?? 'closed'
-      this.newState = init.newState ?? 'closed'
-    }
-  }
-  ;(globalThis as unknown as { ToggleEvent: typeof ToggleEventShim }).ToggleEvent = ToggleEventShim
-}
-
-if (!('popover' in HTMLElement.prototype)) {
-  const popoverOpen = new WeakMap<HTMLElement, boolean>()
-
-  Object.defineProperty(HTMLElement.prototype, 'popover', {
-    configurable: true,
-    get(this: HTMLElement) {
-      return this.getAttribute('popover')
-    },
-    set(this: HTMLElement, value: string | null) {
-      value === null ? this.removeAttribute('popover') : this.setAttribute('popover', String(value))
-    },
-  })
-
-  HTMLElement.prototype.showPopover = function (this: HTMLElement) {
-    if (popoverOpen.get(this)) {
-      return
-    }
-    this.dispatchEvent(new ToggleEvent('beforetoggle', { oldState: 'closed', newState: 'open' }))
-    popoverOpen.set(this, true)
-    this.dispatchEvent(new ToggleEvent('toggle', { oldState: 'closed', newState: 'open' }))
-  }
-
-  HTMLElement.prototype.hidePopover = function (this: HTMLElement) {
-    if (!popoverOpen.get(this)) {
-      return
-    }
-    this.dispatchEvent(new ToggleEvent('beforetoggle', { oldState: 'open', newState: 'closed' }))
-    popoverOpen.set(this, false)
-    this.dispatchEvent(new ToggleEvent('toggle', { oldState: 'open', newState: 'closed' }))
-  }
-
-  HTMLElement.prototype.togglePopover = function (this: HTMLElement) {
-    popoverOpen.get(this) ? this.hidePopover() : this.showPopover()
-  }
-}
 
 window.BASE_URL = 'http://test/'
 window.MAILER_CONFIGURED = true
